requirements.txt
python-telegram-bot==13.15
pydantic==1.10.2
httpx==0.23.1
